+++ download_zone_file.sh
@@ -1,3 +1,3 @@
 #!/usr/bin/env bash
 
-cargo run --package tools --release --bin download_zone_file --manifest-path ./tools/Cargo.toml
+RUSTFLAGS="-A dead_code -A unused_variables -A unused_mut -A unused_imports" cargo run --package tools --release --bin download_raw_zone_file --manifest-path ./tools/Cargo.toml
